Convert variables.

+++ src/ZoneTree/Core/ZoneTreeMetaWAL.cs
@@ -50,7 +50,7 @@ public sealed class ZoneTreeMetaWAL<TKey, TValue> : IDisposable
 
     }
 
-    public void EnqueueReadOnlySegment(int segmentId)
+    public void EnqueueReadOnlySegment(long segmentId)
     {
         var record = new MetaWalRecord
         {
@@ -60,7 +60,7 @@ public sealed class ZoneTreeMetaWAL<TKey, TValue> : IDisposable
         AppendRecord(record);
     }
 
-    public void DequeueReadOnlySegment(int segmentId)
+    public void DequeueReadOnlySegment(long segmentId)
     {
         var record = new MetaWalRecord
         {
@@ -70,7 +70,7 @@ public sealed class ZoneTreeMetaWAL<TKey, TValue> : IDisposable
         AppendRecord(record);
     }
 
-    public void NewSegmentZero(int segmentId)
+    public void NewSegmentZero(long segmentId)
     {
         var record = new MetaWalRecord
         {
@@ -80,7 +80,7 @@ public sealed class ZoneTreeMetaWAL<TKey, TValue> : IDisposable
         AppendRecord(record);
     }
 
-    public void NewDiskSegment(int segmentId)
+    public void NewDiskSegment(long segmentId)
     {
         var record = new MetaWalRecord
         {
