+++ src/ZoneTree/Segments/Disk/FileRandomAccessDevice.cs
@@ -23,7 +23,7 @@ public sealed class FileRandomAccessDevice : IRandomAccessDevice
 
     public FileRandomAccessDevice(
         IFileStreamProvider fileStreamProvider,
-        int segmentId,
+        long segmentId,
         string category,
         IRandomAccessDeviceManager randomDeviceManager,
         string filePath, bool writable, int fileIOBufferSize = 4096)
